removal of parallel reading of eureka endpoint
