+++ api-catalog-services/src/main/java/org/zowe/apiml/apicatalog/controllers/handlers/ApiCatalogControllerExceptionHandler.java
@@ -10,7 +10,7 @@
 
 package org.zowe.apiml.apicatalog.controllers.handlers;
 
-import org.zowe.apiml.apicatalog.controllers.api.ApiCatalogController;
+import org.zowe.apiml.apicatalog.controllers.api.ServicesController;
 import org.zowe.apiml.apicatalog.exceptions.ContainerStatusRetrievalThrowable;
 import org.zowe.apiml.message.api.ApiMessageView;
 import org.zowe.apiml.message.core.Message;
@@ -24,7 +24,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 /**
  * This class creates responses for exceptional behavior of the ApiCatalogController
  */
-@ControllerAdvice(assignableTypes = {ApiCatalogController.class})
+@ControllerAdvice(assignableTypes = {ServicesController.class})
 @RequiredArgsConstructor
 public class ApiCatalogControllerExceptionHandler {
     private final MessageService messageService;
