+++ api-catalog-services/src/main/java/org/zowe/apiml/apicatalog/controllers/handlers/CatalogApiDocControllerExceptionHandler.java
@@ -10,7 +10,7 @@
 
 package org.zowe.apiml.apicatalog.controllers.handlers;
 
-import org.zowe.apiml.apicatalog.controllers.api.CatalogApiDocController;
+import org.zowe.apiml.apicatalog.controllers.api.ApiDocController;
 import org.zowe.apiml.apicatalog.services.status.model.ApiDocNotFoundException;
 import org.zowe.apiml.apicatalog.services.status.model.ServiceNotFoundException;
 import org.zowe.apiml.message.api.ApiMessageView;
@@ -25,7 +25,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 /**
  * This class creates responses for exceptional behavior of the CatalogApiDocController
  */
-@ControllerAdvice(assignableTypes = {CatalogApiDocController.class})
+@ControllerAdvice(assignableTypes = {ApiDocController.class})
 @RequiredArgsConstructor
 public class CatalogApiDocControllerExceptionHandler {
     private final MessageService messageService;
