+++ api-catalog-services/src/test/java/org/zowe/apiml/apicatalog/controllers/api/ServicesControllerContainerRetrievalTest.java
@@ -28,22 +28,22 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @ExtendWith(SpringExtension.class)
-@WebMvcTest(controllers = {CatalogApiDocController.class},
+@WebMvcTest(controllers = {ServicesController.class},
     excludeFilters = { @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, value = WebSecurityConfigurer.class) },
     excludeAutoConfiguration = { SecurityAutoConfiguration.class}
 )
-@ContextConfiguration(classes = CatalogApiDocControllerServiceNotFoundTestContextConfiguration.class)
-class CatalogApiDocControllerServiceNotFoundTest {
+@ContextConfiguration(classes = ApiCatalogControllerContainerRetrievalTestContextConfiguration.class)
+class ServicesControllerContainerRetrievalTest {
 
     @Autowired
     private MockMvc mockMvc;
 
     @Test
-    void getApiDocForServiceDown() throws Exception {
-        this.mockMvc.perform(get("/apidoc/service1/v1"))
-            .andExpect(status().isNotFound())
-            .andExpect(jsonPath("$.messages[?(@.messageNumber == 'ZWEAC706E')].messageContent",
-                hasItem("Service not located, API Documentation not retrieved, The service is running.")));
+    void getContainers() throws Exception {
+        this.mockMvc.perform(get("/containers"))
+            .andExpect(status().is5xxServerError())
+            .andExpect(jsonPath("$.messages[?(@.messageNumber == 'ZWEAC104E')].messageContent",
+                hasItem("Could not retrieve container statuses, java.lang.NullPointerException")));
     }
